utils_sys: moved update_boot_option from utils_test
Moved update_boot_option from utils_test to utils_sys so that
__init__.py is not as long.
